pyproject.toml
[project]
name = "msagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=1.0.0",
    "azure-identity>=1.25.1",
    "python-dotenv>=1.1.0",
]
